requirements.txt
torch>=2.1.0
transformers>=4.44.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
datasets>=2.20.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
sentencepiece>=0.2.0
protobuf>=4.25.0
